requirements.txt
colorama==0.4.6
dnspython==2.7.0
prettytable==3.11.0
python-dateutil==2.9.0.post0
python-whois==0.9.4
six==1.16.0
wcwidth==0.2.13
